move plot_coverage() and align_and_plot_coverage() to reports.py
move plot_coverage() and align_and_plot_coverage() to reports.py, along
with their respective parser functions
